Consolidate client data types cont'd
